completed refactoring for package location changes
